refactor: bring out latex component to reduce rendering
